add refreshModel to namespace query param
There is a bug in the Ember Router service that does not preserve query parameters
during a partial transition which is what happens when we forward a query parameter
using transitionTo or LinkTo. That initializes a new set of queryParameters and since
we're inheriting from a sibling route, the new navigated to controller and route have
no knowledge of that previous query parameter. To work around this, we rely on adding
refreshModel to the queryParameter because refreshModel calls Router.refresh which
maps all query params from the previous state into the refresh. Since we're relying on
this bug in the API (Hyrum's Law), we'll follow the Beyonce rule and test the dependency
on the API in an acceptance test (if you like it, then you shoulda put a test on it). If
we update Ember and this test fails, its because Ember fixed this Router bug.